SPXSubscriptionManager: Add detailed completion status information.
Change the completion handler of
`purchaseSubscriptionProduct:completionHandler:` to receive
`BZRReceiptSubscriptionInfo` on success or `NSError` on failure.
Change the completion handler of
`restorePurchasesWithCompletionHandler:` to receive `BZRReceiptInfo`
on success or `NSError` on failure.

For both methods on cancellation the completion handler will be invoked
with first parameter set to `nil` and `error.errorCode` will be
`BZRErrorCodeOperationCancelled`. In that case failure alert will not
be presented.