PTNDataAssetCache: Workaround for caching object on same URL.
Any cached item can use any URL when caching. When, for example, an
image was cached with a URL and an album was cached on the same URL
only the last object was cached.

Before this commit when fetching an image on a URL where an album was
cached, a crash occured. The workaround prevents the cache but keeps
the cache with the wrong info.