LTGPUResource: add method for instant disposal.
Since a GPU resource lifetime is currently bound to the ObjC object that
wraps the resource, and it's impossible to force an ObjC object to
deallocate immediately, it is sometimes desireable to have a way to
explicitly dispose the underlying resource without waiting for the
object to deallocate.

This method should be used with caution, as the wrapping object
continues to be alive and can accept messages. Therefore, any side
effects or return values by additional messages sent to the object are
considered undefined behavior.