BZRCachedReceiptValidationStatusProvider: fix flaky test.
The test created a partial mock on an object inside an autorelease
pool, and started an asynchronous operation on it. The underlying
instance of the partial mock was deallocated after exiting the
autorelease pool, which resulted in undefined behavior in the operation
that continued to execute. The partial mock was replaced with a regular
mock.