PNKConcatenation: drop feature channel counts from initializer.
The current implementation of the Concatenation kernel requires
providing feature channel counts of the input images on the kernel
initialization. This requirement makes it impossible to use this kernel
in the framework of CoreML Network Compiler, as the CoreML data format
does not provide feature channel counts for Concatenation. The PR
overcomes this issue by dropping the feature channel count parameters
from the kernel initializer. This change effectively permits for images
with arbitrary feature channel counts to be encoded by the kernel. It
bears a 4% penalty on the elapsed time because some calculations have
been shifted from the compute state creation stage to the running stage.