BZRStore: Fix product eligibility check.
When checking if the user is eligible for acquiring some product via
subscription we check whether the active subscription enables the
requested product. Before this fix, in case there's no product in the
product list for the currently active subscription we consider this as
if the user is not eligible for the requested product no matter which
product is it. After this fix, in the same scenario we treat the
subscription as if it enables all (non-subscription) products.

This is required for working with multi-app subscription since it is
common with multi-app subscription that the subscription product does
not exist in the product list as it is actually originating from a
different app.