Bazaar: fix multi-app receipt validation in case of failure.
Until now, when validating using the multi-app receipt validation, if
the receipt validation failed, the last successful receipt validation
status was not used instead. This can cause a scenario where there was a
temporary failure validating the receipt validation status and the user
doesn't have access to its subscription, until he restarts the app.
Moreover, this was the behaviour when the single-app validation took
place.