LTDynamicQuadDrawer: Disallow non-trivial projections.
Originally, the LTDynamicQuadDrawer was implemented for usage without a
custom projection matrix. In particular, it assumes the used projection
to be orthographic and perpendicular to the plane in which the rendered
quads reside. Unfortunately, this assumption was neither documented in
the header file of the class nor enforced by the implementation. This
commit mitigates this lack of documentation and adds a corresponding
assertion.

Vertex shaders requiring a custom projection or, conceptually, a
modelview  matrix, need to implement it by adding such matrix. If
necessary due to the specific projection/modelview matrix used, the
correct functioning of the shader must be ensured by zeroing out the
z-coordinate of the vertex position before multiplying the position with
the matrices.