SPXPagingView: Fix scrolling behavior when the velocity is low.
The target index depends on the swipe velocity. If the velocity is
positive the direction will be right (ceil) and if negative the
the direction will be left (floor).
But when the velocity was zero, the flooring still happened and the
direction was always left. The desired behavior in this case is to
take the closest index by rounding.