PNKComputeDispatch: replace MTLTexture in APIs with MPSImage.
All compute dispatch APIs receiving an array of MTLTextures have been
replaced with APIs receiving two arrays of MPSImages. These arrays are
called inputImages and outputImages. This way we can decrement the
readCount property of all input images of type MPSTemporaryImage in this
low-level API instead of doing this in each kernel separately.