DVNBrushModelV1: Refine specification and use additional image URLs.
Refines the specification constituted by the DVNBrushModelV1 class. The
final brush tip is now assumed to be computed based on two images, the
so-called source image, and the so-called mask image.
In addition to the aforementioned refinement, an additional image URL
for the edge avoidance image is added and lt::Interval is used where
applicable.