Bazaar: Remove `pendingRenewalInfo` post deserialization validation.
This commit removes the post-deserialization model validation of
`BZRSubscriptionPendingRenewalInfo`. The reason for this removal is that
it appears that the pending renewal can take many unexpected formats.
For example the `willAutoRenew` flag may be `NO` and yet the
`autoRenewProductId` will be non-null, on the other hand the
`expirationReason` may be `nil` even if `willAutoRenew` is `NO`.

Instead of validating the combination of the model values we accept
Apple's response as is, and now apps should be checking the values of
this model and present alerts accordingly.