INTAnalytricksTransformerBlocks: Enrich events with unique event id.
The analytricks metadata enricher was using the eventID of the metadata
object that relates to the low level event that triggeres the creation
of high level event. This method resulted in a bug where if many high
level events were created for one low level, they would all have the
same event id. This fix creates an event id at the enrichment.