WFVideoViewTests: Fail tests on error from player.
Flakiness of tests that practice playback in WFVideoViewTests in CI has
been observed. It is probably because at some conditions an error is
reported from the player. This (temporary) commit is in order to get the
details of this error from CI and understand what causes it and how to
fix this flakiness.