Bazaar: add fake replacement to SKPaymentQueue.
There was a problem when testing `BZRPaymentQueue` - If we were to use a
mock object as the `SKPaymentQueue`, it would result in a retain cycle.
`BZRPaymentQueue` passed itself to the mock as an observer, which
resulted in the mock holding `BZRPaymentQueue` strongly (the correct
behaviour when mocks methods are called with arguments). This resulted
in a retain cycle because `BZRPaymentQueue` itself holds the mock
strongly.  That's why we had to resort to using an actual
`SKPaymentQueue`.
Instead, an internal payment queue that has exactly the same methods as
`SKPaymentQueue` was added so that we will be able to use it in tests
and not the real `SKPaymentQueue`.