LTInterval: Add method for clamping interval to another interval.
Adds a method for clamping an interval to another interval.
The returned value is one of the following:
a) An empty optional if the interval to clamp to is empty.
b) The intersection of the two involved intervals, if the intersection
   is non-empty.
c) The non-empty interval consisting of the single value a, s.t. a is an
   arbitrary value contained by this interval clamped to the given
   interval.