Fiber: remove request from error when receiving failed response.
Whenever a response with failure was received, the request was
attached to the `NSError`. Although this may be more informative, we
assume that the response give enough information for us to know why the
HTTP request failed, whether it has something to do with the request or
not.
In one use case, the request to Validatricks contained the whole receipt
data which is occasionally very large and causes the failure event to be
thrown away by our analytics tools.