PTNPhotoKitAssetManager: fetch image data off the main thread.
`PHImageManager`'s fetch methods call their completion blocks on the
main thread, but the returned assets may be used on a different threads.
Moreovers, this may cause deadlocks if another subsystem, such as
AVFoundation, requires an asset synchronously on the main thread.

To mitigate the issue, image data is now fetched using
`PHAssetResourceManager`, which doesn't call its callbacks on the main
thread. This requires some extra logic that decides which resource to
fetch, which was derived experimentally by comparing `PHImageManager`'s
returned data and the appropriate resource that backs it.

Note: the new logic is not 1:1 compatible with the previous one.
The previous fetch used the `PHImageRequestOptionsVersionCurrent` option
which may not retrieve a high-fidelity version of the image, while the
new logic may retrieve it in some scenarios. An example for this is
fetching screenshots that are saved in PNG, but `PHImageManager` returns
them as JPEGs.