Photons: drop Dropbox source.
The source is currently not used in any client. Moreover, the currently
linked SDK is no longer supported. If we choose to add this source in
the future, this code should serve as an initial point for integration
with the new SDK, but there's no reason to force apps to link against
it.