SHKTweakCategoryKiosk: remove language tweak.
The tweak used a global value in NSUserDefault to control the language
and this caused problems when the selected language was the device
language.

1. When the selected language was the device language, the value was
set to `nil` which caused random crashed.

2. In implementations where `nil` was not used as the value, but rather
the language of the current locale, when the device language was
changed, the key was not updated by the system for some reason, and
the old value was used, until the tweak ran and rewrote the value of
the new language.

For these reasons, it is unreasonable to control the language in with
this method. Remove the tweak until a better method is found.