Bazaar: move Validatricks models logic into base classes.
In order to simplify the models in Bazaar, Validatricks models are to
be merged into their parents. However, if we were to delete the models
now, the data stored in the secure storage would be invalid. Therefore,
for now we keep the Validatricks classes and move all the code that
deals with Validatricks models to their base classes.