Bazaar: fix serialization problems with NSDate.
In the tests Bazaar serializes `NSDate` properties to and from a double
value representing milliseconds since 1970. Apparently, there is a bug
when converting an `NSDate` to and from `NSTimeInterval` using
`timeIntervalSince1970`: The result date is not always equal to the
start date. This happens when double values with high precision (such
as those created with `[NSDate date]` and
`[NSDate dateWithTimeIntervalSinceNow]`) are given as input, since
internally `timeIntervalSince1970` adds it to another double value,
which results in a value that occasionally can't be represented in a
double, which causes a small rounding error.
Since it happens only when the function is given dates created from
functions such as `[NSDate date]` and
`[NSDate dateWithTimeIntervalSinceNow]`, the tests are changed to use
`dateWithTimeIntervalSince1970:` that is given an integer, to ensure
the conversions are consistent.