RACSignal+Photons: Generalize ptn_skipProgress.
The skip progress operator was intended to enable easy opt-out of
progress information where it is not relevant. It performed type checks
back when there was only one valid type. But now many signals in Photons
send progress wrapping various results. This commit removes the
verification of the result's type, enabling usage of the operator for
other progress signals provided by Photons.