Pinky: use uint read and write variants to work around rdar://39206631.
An issue that was surfaced in iOS 11.3 causes texture misreads when
using the `ushort` variants of `texture2d_array`. This change works
around it by using coordinates with type `uint` instead. Since the
buffers passed to the GPU are extremely small and the variables inside
the kernel will be assigned to registers anyway, the performance
implication of this change is expected to be negligible.