LTAppIntegrity: validate file magic to be of a Mach-O header type.
The app's binary can start immediately with a Mach-O header, or can
start with a fat binary header that contains multiple Mach-O slices.
Since this case is not valid in production at the moment (because we
only ship thin arm64 binaries), we avoid parsing fat binaries and only
accept thin files. Additionally, this validation only approves Mach-O
files that are in the host format (i.e. we're not accepting MH_CIGAM_*
magics) since the code is not swapping the appropriate structs in this
case.