PTNOceanAssetManager: choose image size according to resizing strategy.
Before this commit, the specific image size to return was chosen by
minimizing the euclidean distance between the image size and the size
provided by the resizing strategy.

Using the above method is simplistic and does not fit all resizing
strategies. Fix this by using the minimal size that bounds the size
required by the resizing strategy.