DVNBrushStroke(Specification/Data): Initial commit.
Adds DVNBrushStrokeSpecification, a value class representing a brush
stroke, as defined by the DVNBrushModel class. In particular, the value
class holds the information required to construct a spline from an
LTControlPointModel and perform the rendering of the brush stroke along
the spline, according to a DVNBrushRenderModel.

In addition, adds DVNBrushStrokeData a class holding both a light-weight
DVNBrushStrokeSpecification object and the corresponding texture
mapping.