DVNBrushStrokePainter: Initial commit.
Adds DVNBrushStrokePainter, an object representing a "painter" of "brush
strokes", i.e. an entity which draws with a "brush" along a spline onto
a "canvas", yielding a visual brush stroke. The brush is described by a
DVNBrushModel, the spline consists of a sequence of LTSplineControlPoint
objects, a so-called spline control point sequence, and the canvas is
either an LTTexture or a render target assumed to be bound during the
rendering. For every spline control point sequence, a different brush
and canvas can be used.