Bazaar: Replace SSL certificate pinning with public key pinning.
This is required since SSL certificate pinning verifies the entire
server certificate file. While probably most secure this is not a good
approach for SSL pinning because the certificate on the server is being
replaced with a new one when the old one is about to be expired. So
users that only have the old certificate pinned will not be able to
connect to the server once its certificate is renewed. Instead pinning
only the certificate's public key may be better since renewing the
certificate keeps the same private/public key pair, hence users pinning
only the public key will still be able to connect even after renewal.