Revert "PTNOceanAssetManager: choose image size according to resizing strategy."
This reverts commit da1057e249bfbb2436a9073c0adc6236177bb7ed.

Selection of the most appropriate image size to download based on a
resizing strategy was altered in a previous commit to use the
-inputSizeBoundedBySize: method. Unfortunately this method was
intended for a scenario where the possible sizes are of a fixed array,
and the actual image sizes are only known to be aspect fitted to those
sizes (Dropbox API). Ocean's API gives actual asset sizes which
simplifies this decision.

The main issue is that under the Dropbox API assumptions, limiting the
asset size of aspect filled resizing strategies is impossible, since
the aspect ratio of the underlying image can always cause the bounding
size to not be enough. To make things clearer with an example -
Dropbox gave a constant set of possible fetch sizes (for example (100,
100) and (200, 200)) and then promised to provide images that are
smaller than the requested size in both dimensions. So a resizing
strategy of aspect fill to (20, 20) couldn't settle for (100, 100)
because if the underlying image is of size (10, 100), the actual size
you want is (20, 200).

Here this isn't the case, we get the actual size and all we need to
verify is that its larger than the aspect fill strategy's size in both
dimensions.

The problem then was that for aspect fill strategies this approach
always chooses the largest size available - a necessity for the
Dropbox API, a waste here. This resulted in high bandwidth usage and
depending on the user's network, noticeably slower performance during
scrolling.

The solution that existed before this change isn't correct either, its
more of a heuristic that needs to be amended in the future, simply
with a more correct method. For now I believe we should opt for a
quick fix though as this issue is already found in released versions.

I'll be the first to admit that this was a pretty complicated API to
understand.