BZRReceiptValidationErrorDeviceIDMismatch: Add JSON transformation.
This error should be sent by the server in case the receipt file sent
for validation was actually generated for a different device. Currently
the Validatricks server does not verify this value but it will in the
future in order to fight piracy.
Adding JSON transformation to this value is also needed in order to
make the value-transformer reversible.