BZRBillingPeriod+Shopix: Parse billing period from the last group.
This commit fixes a bug caused due to wrong parsing.
For example: for the product ID `com.spx.V1.1M.SA_1Y.SA` where the tier
group is `V1.1M.SA` and the product group is `1Y.SA` the billing period
should be `1Y`, but splitting by `.` resulted in the component `SA_1Y`
which is not equal to `1Y` as expected.

The solution is to look for the period string in the product group only.