NSErrorCodes+LTKit: add support for subsystem ID.
As the number of error codes defined in our projects grew, we noticed
two issues with the current way errors are defined:
1. The number of supported errors is limited by 40 due to the way the
   error code definition macro is implemented.
2. Errors are defined in a single file per product, where it's may be
   more comfortable to define them per subsystem instead.

This commit adds an option to specify subsystem ID for each error code,
together with the existing product ID. This is done by allocating
another 4 bits in the error code for the subsystem ID, right below the
product ID. This option is introduced in the
`LTErrorCodesWithSubsystemDeclare` and
`LTErrorCodesWithSubsystemImplement` macros.

The changes are backward-compatible, and existing error codes values are
preserved by setting a default subsystem ID of 0 for users of the
existing macro.