NSValueTransformer+Photons: accept nil in URL value transformer.
nil was already accepted (in forward transformer), but resulted in
LogError. In reverse direction, nil triggers assert.

With this change, nil is accepted in both directions. This is the sane,
symmetrical thing to do.

Without this change, video search responses from Ocean generate endless
LogErrors, which in turn generate non-fatal crashes in Crashlytics, and
slow everything down.