BZRAppStoreLocaleProvider: always set App Store locale after fetch.
The App Store locale property was previously set only when there was a
difference between the current App Store locale and the one that was
fetched. However in one use case `BZRStore` needs to wait for the
fetched one - which is not possible if we don't set the property after
the fetch finished.