LTGuidedFilterCoefficientsProcessor: accept non uniform kernels.
Allow coefficients textures to have a very different aspect ratio
comparing to the input texture so no exception will be raised for
corner case inputs.

Since Core Image doesn't support separate width and height of kernel for
box filter, the internal blur result will be only meaningful only if the
aspect ratio of the input is close to the aspect ratio of the
coefficients textures.