BZRBillingPeriod+Shopix: fix getting period of product with trial.
The previous implementation assumed that the last component always
contains the subscription period, however it is not true because the
discount specification can come after it (like trial for example).
Therefore it is assumed that if the format of the product is the new
format, the third component always contains the subscription period.