Warehouse: Change user data from property list to NSData.
The type of the user data and its serialization is not in the scope of
the storage library. Use NSData as user data type to avoid posing
limitations on the user data.

If the user data is a dictionary, the application can for example
create an NSData extension that will contain an initializer that will
serialize a given dictionary and set the serialized data to be the
content of the created object, and a method to deserialize the data to
a dictionary (and store it as an associated object for next calls).