lt_xcodebuild: Add workaround to crash no UBSAN issues.
Other sanitizers crash when an issue is met, so a test case is failed
when, for example, an TSAN issue is found. Until now this was solved
by changing `xcpretty` to consider UBSan issues as failures.

This commit forces a crash when a UBSan issue is found, removing the
need for a custom `xcpretty` build.