Warehouse: Fix static analysis warning.
The warning: "Method accepting NSError** should have a non-void return
value to indicate whether or not an error occurred" raised by static
analyzer for the internal methods that map errors. They are now inlined
in the methods that use them in order to avoid the non standard usage
of NSError as output parameter that doesn't imply an error has occurred
in the method.