Warehouse: Add project and storage size calculations.
The size property of the WHSProjectSnapshot is not the right place for
the project size on disk because the project snapshot should be valid
until the next update but the size might change if the application
makes changes in the assetsURL.

In addition, the value of this property was the size of the project
directory itself, without calculating the size of the actual files in
it.

This commit adds two methods to WHSProjectStorage. One to calculate
the size of a project, and one to calculate the size of the entire
storage. The size, of course, includes all the files and not only the
directory itself.