WFFloatView: Change calculation of visual effect alpha value.
The alpha value of the visual effect when the content is dragged
was calculated so that it will grow linearly from 0 when on an
anchor to 1 when on the horizontally closest dock. It is now changed
to start showing the visual effect only when a certain percent of
the content is hidden and not when leaving the anchor.