Bazaar: revert premature invalidation of receipts.
Following the decision to increase the receipt time to live, this commit
revalidates the receipts that were invalidated due to being disconnected
from Validatricks server, so that the users' subscriptions will be valid
again.

This does not interfere with the periodic validation and the receipt
invalidation. Meaning, they will continue to operate as before (but with
the extended time to live).