Bazaar: Increase cached receipt TTL to 60 days.
Previously we used to invalidate the cached receipt validation status
if more than 14 days had passed since the last successful receipt
validation. The rationale behind this was to be up-to-date with the
subscription status. In particular, we want Bazaar to know when users
cancel their subscription, and in response prevent them from using it.
In addition users who turn off internet connection could theoretically
stay with an active subscription forever with no need to renew it.

Additionally the receipt validation interval was also set for 14 days.
As a result legitimate users with active subscription would be blocked
if 14 days had passed since their last validation and on the 15th day
they had some connectivity issues and couldn't perform the validation.
The problem is that the invalidation of the cached receipt and the
re-validation happens on the same day, which has a great potential to
affect legitimate users.

Increasing the cached receipt validation status TTL to 60 days together
with changing the re-validation interval to something smaller than 60
days gives Bazaar more opportunities to complete a successful validation
before the cache is invalidated. This should tremendously reduce the
number of users that don't have access due to the above mentioned
scenario.