LTParameterizedObject: Use ordered set rather than set.
In order to be able to implement optimizations yielding significant
performance gains w.r.t. sampling parameterized objects, this commit
changes the parameterizationKeys collection of the LTParameterizedObject
protocol from set to ordered set and requires the
mappingForParametricValues: method to return an object with a matrix
whose rows are ordered in the same order as the parameterizationKeys.