DVNScatteredGeometryProviderModel: Fix bug causing incorrect tapering.
This commit fixes a bug in DVNScatteredGeometryProviderModel which
causes the tapering to be incorrect. In particular, the tapering was
computed incorrectly for a) strokes whose length is smaller than the sum
of their desired start and end tapering lengths, and b) strokes whose
end piece started at a parametric value smaller than the length of the
desired start tapering.