DVNScatteredGeometryProviderModelSpec: Use numerically stable value.
Before commit [f6301a5ca8e5ad680b21bfa441b51cfbb939a10c], the computed
value of maxDimension equalled 0.20000000298023224. After the commit,
the value equalled 0.2 due to the slight changes in the numerical
robustness of the updated computations. For this reason, the comparison
value is lowered to 0.15 (which would have been a more reasonable
comparison value from the beginning, due to the fact that maxDimension
was so close to 0.2 also before).