BZRReceiptValidationDateProvider: fix computation of validation date.
The next date of validation used to be computed as the earliest of the
validation dates of all apps in the ecosystem, this can a loop if the
user installed an old version of another app in the ecosystem, as the
validation date of this app will never be updated.

In this commit. we change the computation to be dependant on the date of
validation of the currently used \c receiptValidationStatus only, and
ignore other apps.