PNKStyleTransferProcessor: add constraint on the large side of input.
In the current implementation of PNKStyleTransferProcessor we have only
a constraint on the small side of the network input image
(default 1024). As a result - long strips with small side of 1024 or
less are transferred to the CNN as is, without resizing. When the large
size is big enough (4096 or more) this leads to GPU memory overflow. In
this fix we introduce an additional constraint on the large size of the
input image (default 3072) that prevents this crash.