DVNDirectedRectProvider: Correctly handle sequences of single samples.
Fixes a bug which causes DVNDirectedRectProvider to behave incorrectly
in case of sequences of id<LTSampleValues> objects starting with an
id<LTSampleValues> containing a single sample. In particular, the bug
causes the corresponding quad to have size 0. The solution is to buffer
the initial id<LTSampleValues> object, if required, and concatenate it
with the consecutively provided one, in order to reach a minimum number
of two samples.