OBJCObject.py: fix the regexp to stop dropping numbers in names.
A mistake in the regexp caused removal of numbers in the middle of
property names, instead of removing numbers in the beginning of them.
This PR fix the regexp to remove numbers in the beginning.

For example: "level_0_error" is now converted to "level0Error"
and "0level_error" is converted to "levelError".

In the past "level_0_error" was converted to "levelError" and
"0level_error" to "0levelError".

This is clearly broken as an objc property can start contain numbers,
but not in the beggining.