DVNBrushRenderConfigurationProviderV1: Use improved min scale factor.
Rather than only taking into account the minimumTaperingScaleFactor of
DVNBrushModelV1 (which is a multiplicative factor and hence can yield
unpleasantly small brush tips in case that the scale of DVNBrushModelV1
is already small), the maximum of the aforementioned
minimumTaperingScaleFactor value and the appropriate value based on the
scaleRange of the DVNBrushModelV1 object is used. This ensures that the
the size of the brush tips is effectively bounded by the scaleRange.