LTVector: move free clamp functions to member functions.
In C++17, a new `std::clamp` function was introduced. Since this
function relies on the `<` operator which doesn't exist for `LTVector`
and it's illegal to extend the STL in such a way, these methods need
to be either extracted to a different namespace or to be a member
function, which makes sense for now as there are other member functions
on these classes and there's no well-defined namespace for these methods
(yet).