LTCGExtensions: remove `std::clamp` functions.
C++17 introduced a new `std::clamp` function that behaves differently
than the one implemented in LTCGExtensions: it doesn't accept inverted
range (lo > hi), and the result is undefined. Since this means that our
(illegal) extension to STL is now breaking the definition of this
function, we have to remove declarations of it.

Practically, calls to `std::clamp` will still work for `CGFloat`s and
types that implement the `<` operator. However, it is no longer legal to
pass an inverted range to `std::clamp`, so this is a breaking API
change.

Finally, we implement the `CGPoint` clamping function to `CGRect` as a
regular LTCGExtension function named `CGPointClampToRect`, since the
behavior there is completely different than what expected in
`std::clamp`.