LTParameterizedObjectConstructor: Expose (number of) control points.
Since the constructor is the source of truth regarding the control
points used for construction of the parameterized object, it should
publicly expose them. For efficiency, a method returning the number of
control points without the need to create a copy of the control points
array is added as well.