DVNBrushStrokePainter: support spline smoothing.
Added the DVNBrushStrokePainterDelegate an optional -auxiliaryCanvas
method which returns a helper texture to be used by the painter when
splineSmoothness is set in the DVNBrushModel. This canvas is required
for drawing only the smoothed points, which aren't subject to any
changes. This in contrast to spline's trailing points which are mapped
during the smoothing for pleasent look. Later the trailing points are
drawn to the brushStrokeCanvas.