BZRAppStoreLocaleProvider: refactor locale loading signal.
Previously BZRStore observed \c appStoreLocaleProvider.appStoreLocale,
and ignored it's first value. This was because BZRStore assumed that the
first value is the cached value and we wish to use the second value
which is fetched from the product list.

However this is a hidden assumption, based on internal implementation
detail, and it can cause race conditions which causes flakines in tests.

Instead of relying on this assumption, \c BZRAppStoreLocaleProvider
now has a property that is changed to \c YES when the data is loaded
from the product list, and \c BZRStore observes the property.