SPXSubscriptionViewModel: Populate introductory discount properly.
VM is currently only populating the price info upon fetch completion.
Introductory discount of the product is left untouched. So there is no
way to know what is the introductory discount in runtime which is an
issue if one wants to display a compact description of the discount to
the user.