INTDeviceInfo: add device model property.
Before this commit, only the device kind was sent, which is a human
readable version of the device model, something like "iPhone X". This
commit adds the device model, "iPhone9,4" for example.

This is done since unknown models are sent as "Unknown" and the exact
model is not specified. Moving the mapping between the device model to
the query-time enables us to query events even if the mapping did not
exist in the device in the time when the event was sent.