LTPath: remove side effect from -pathWith* initializers.
The private (designated) initializer
 -initWithBaseDirectory:andRelativePath: uses
 -stringByStandardizingPath method which behaves as follows:

> A new string made by removing extraneous path components from the
> receiver.
>
> A new string made by performing the following operations:
>
> * Removing an initial component of "/private/var/automount",
>   "/var/automount", or "/private" from the path, if the result still
>   indicates an existing file or directory (checked by consulting the
>   file system).

Hence LTPath created with a full path: /private/foo, for example, will
result with two unequal LTPath objects depending whether /private/foo
file exist or not.

In test a LTPath p is created, then a file f is stored at path p.
Lastly, LTPath corresponding with f is different than p.
This effect causes tests to fail when running on a device.

Another issue with using -stringByStandardizingPath is it does I/O (to
figure out if the file exists), which is a unexpected from LTPath
(value class) initalizer.