LTIterativeProcessingStrategy: fix iteration placement bug.
`LTNextIterationPlacement` returned by `iterationStarted` wasn't
calculated correctly i.e. the alternation algorithm
between *intermediate* and *output* returned incorrect values for odd
iteration numbers.

The introduced algorithm takes into an account the current iteration
and the desired number of iterations for an output. Number of steps is
the difference between the up mentioned numbers. If the number of
steps is odd, set the target to point to output, otherwise target
points to intermeditate.