LTTestUtils+LTEngine: add LTAttachImagesToCurrentTest() method.
This method allow attaching named images to current test's
results. Additionally remove the old infrastructure which wrote the
compared matrices into /tmp directory. The implemented flow works on
both: a simulator and a device.

Exposed the internal LTWriteMatAsPNG method which allows writing
cv::Mat as PNG images to a path.