LTCompressionFormatSpec: make isSupported property a simulator test.
This test checks that each compression format is supported on the
platform under test. Support of HEIC compression format is dependent
on: iOS version, wether it's simulator or device and the device type.

The cross product of iOS version, platform type and device type is big
and testing it has low value. Hence limit the test to run only on
simulator for sanity only.